users can now attend just by clicking attend

+++ backend/controllers/AttendanceController.js
@@ -3,6 +3,7 @@ var models  = require('../models');
 var Sequelize = require('sequelize');
 var Attendance = models.Attendance;
 var helper = require("../helpers/helper.js");
+var userService = require('../services/UserService.js');
 
 module.exports = {
 	findOne: function (req, res) {
@@ -36,30 +37,50 @@ module.exports = {
 		console.log(attendanceToAdd.event)
 		console.log(attendanceToAdd)
 
-		Attendance.create({
-			name: attendanceToAdd.name,
-			email: attendanceToAdd.email,
-			comment: attendanceToAdd.comment,
-			EventId: attendanceToAdd.event.id
-		}).then(function(attendance){
-			res.send(attendance)
-		}).catch(function(err){
+        var success = function(user){
+            Attendance.create({
+                name: attendanceToAdd.name,
+                email: attendanceToAdd.email,
+                comment: attendanceToAdd.comment,
+                EventId: attendanceToAdd.event.id,
+                userId: user.id
+            }).then(function(attendance){
+                res.send(attendance)
+            }).catch(function(err){
+                helper.sendErr(res, 400, err);
+            });
+        }
+        var error = function(err){
             helper.sendErr(res, 400, err);
-        });
+        }
+        userService.getLoggedInUser(req, res, success, error);
 	},
 
     //TODO: add check that the user is signed in and the creator of the event
     delete: function(req, res){
-        var attendanceId = req.params.id;
-        models.Attendance.destroy({
-            where: {
-            id: attendanceId
-            }
-        }).then(function(){
-            res.status(200).send();
-        }).catch(function(err){
-            helper.sendErr(res, 400, err);
-        });
+        var success = function(user){
+            var attendanceId = req.params.id;
+            Attendance.findOne({
+                where: { id: attendanceId }
+            }).then(function (attendance) {
+                if( user.id === attendance.userId ){
+                    models.Attendance.destroy({
+                        where: {
+                            id: attendanceId
+                        }
+                    }).then(function(){
+                        res.status(200).send();
+                    }).catch(function(err){
+                        helper.sendErr(res, 400, err);
+                    });
+                }else{
+                    helper.sendErr(res, 400, err);  //not authorized
+                }
+            }).catch(function(err){
+                helper.sendErr(res, 400, err); 
+            });
+        }
+        userService.getLoggedInUser(req, res, success, error);
     }
 
  /*	update: function (req, res) {
